+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 实现预设调度器。
+  - com.dwarfeng.statistics.impl.handler.dispatcher.DrainDispatcher。
+
 - 实现预设驱动器。
   - com.dwarfeng.statistics.impl.handler.driver.CronDriverProvider。
   - com.dwarfeng.statistics.impl.handler.driver.CronDriverSupporter。
@@ -20,6 +23,7 @@
 
 - 实现核心机制。
   - 驱动机制。
+  - 调度机制。
 
 - 定义预设服务。
   - com.dwarfeng.statistics.impl.service.EnabledDriverInfoLookupServiceImpl。